workers/separation/pyproject.toml
[project]
name = "tidaldub-separation"
version = "0.1.0"
description = "Audio source separation worker using Demucs"
requires-python = ">=3.12,<3.15"

dependencies = [
    # PyTorch ecosystem - latest with CUDA 12.4
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    
    # Audio separation
    "demucs>=4.0.1",
    
    # Numerical
    "numpy>=2.2.2",
    "scipy>=1.15.1",
    
    # Progress
    "tqdm>=4.67.1",
    
    # Config
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.uv]
# GPU optimizations for RTX 5070 (8GB VRAM)
# Demucs htdemucs uses ~4-6GB VRAM

[tool.uv.sources]
torch = { index = "pytorch-cuda" }
torchaudio = { index = "pytorch-cuda" }

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
